cloud_runs/python/econ_export/pyproject.toml
[tool.poetry]
name = "econ_export"
version = "0.1.0"
description = ""
authors = ["Maikel Ruiz <m.ruiz@combocurve.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
dnspython = "^1.16"
fastapi = "^0.89.1"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
mongoengine = "0.20.0"
pusher = "3.0"
google-cloud-storage = "1.32.0"
pymongo = "3.11"
pandas = "1.3.5"
numpy = "1.24.2"
poethepoet = "^0.18.1"
xlrd = "2.0.0"
pyarrow="^10.0.0"
polars = "^0.15.18"
joblib = "^1.2.0"
google-cloud-bigquery = {extras = ["bqstorage", "pandas"], version = "^3.4.1"}
google-cloud-logging = "^2.6"
google-resumable-media = "^1.3"
scipy = "1.10.0"
Flask = "^2.0"
Flask-Cors = "^3.0"
jsonschema = "^4.17.3"
grpc-interceptor = "^0.15.1"
google-cloud-secret-manager = "^2.16.2"
launchdarkly-server-sdk = "^8.1.4"

[tool.poe.tasks]
test-unit = 'pytest -m unittest --disable-warnings'
test-integ = 'pytest -m integtest'
test-bench = 'pytest -m benchmark --benchmark-json benchmark-output.json'
lint = "flake8"


[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
yapf = "^0.32.0"
toml = "^0.10.2"
freezegun = "^1.2.0"
pytest = "^7.2.1"
pytest-mock = "^3.10.0"
pytest-watch = "^4.2.0"
pytest-testmon = "^1.4.5"
pytest-cov = "^4.0.0"
python-dotenv = {extras = ["cli"], version = "^0.21.1"}
deepdiff = "^5.8.1"
pydantic = "^1.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
